+++ src/Uno.UWP/Security/Authentication/Web/WebAuthenticationBroker.Android.cs
@@ -6,9 +6,6 @@ namespace Windows.Security.Authentication.Web
 {
 	public static partial class WebAuthenticationBroker
 	{
-		public static bool OnResume(Intent? intent=null)
-		{
-			return (_authenticationBrokerProvider as WebAuthenticationBrokerProvider)?.OnResumeCallback(intent) ?? false;
-		}
+		internal static bool OnResume(Intent? intent = null) => (_authenticationBrokerProvider as WebAuthenticationBrokerProvider)?.OnResumeCallback(intent) ?? false;
 	}
 }
